requirements.txt
openai>=1.30.0
tweepy>=4.14.0
feedparser>=6.0.10
httpx>=0.27.0
trafilatura>=1.7.0
beautifulsoup4>=4.12.3
pydantic>=2.7.0
python-dotenv>=1.0.1
tldextract>=5.1.2
schedule>=1.2.1
pytz>=2024.1
tenacity>=8.5.0
PyYAML>=6.0.2

